get weather data from api to home fragment

+++ app/src/main/java/com/example/android/weatherapp/model/data/Current.kt
@@ -24,5 +24,5 @@ data class Current(
     @SerializedName("wind_deg")
     val windDeg: Int,
     @SerializedName("wind_speed")
-    val windSpeed: Int
+    val windSpeed: Double
 )
